drivers: npmx: shell: move adc code to the separate file
