+++ drivers/npmx/CMakeLists.txt
@@ -46,6 +46,8 @@ zephyr_library_sources(npmx_driver.c)
 if(CONFIG_NPMX_SHELL)
     zephyr_library_sources(npmx_shell.c)
     zephyr_library_sources(shell/shell_common.c)
+
+    zephyr_library_sources(shell/adc.c)
 endif()
 
 # Get npmx include file
